requirements.txt
azure-identity==1.12.0
Flask==2.3.2
openai==0.27.8
azure-storage-blob==12.14.1
python-dotenv==1.0.0
langchain==0.0.325
azure-ai-formrecognizer==3.2.0
requests==2.31.0
tiktoken==0.2.0
azure-storage-queue==12.5.0
beautifulsoup4==4.12.0
fake-useragent==1.1.3
chardet==5.1.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.4.0b8
opencensus-ext-azure==1.1.9
azure-ai-contentsafety==1.0.0b1
python-docx==0.8.11
azure.kusto.data